requirements.txt
streamlit==1.25.0
pandas==1.5.3
plotly==5.9.0
inflection==0.5.1
folium==0.14.0
Pillow==9.4.0
plotly-express==0.4.1
streamlit-folium==0.13.0
